Separate handler so it's easier to add multiple LLM providers
